feat(cli): add module entrypoint
