+++ AGENTS.md
@@ -23,7 +23,7 @@
 
 #### Точка входа
 - `app/main.py` — запуск графического приложения и создание главного окна.
-- `app/cli.py` — консольный интерфейс: работа с требованиями и проверка настроек.
+- `app/cli/main.py` — консольный интерфейс: работа с требованиями и проверка настроек (запуск через `python3 -m app.cli`).
 
 #### Настройки и конфигурация
 - `app/settings.py` — Pydantic‑модели `LLMSettings`, `MCPSettings`, `UISettings` и агрегирующий `AppSettings`.
