+++ app/cli/__init__.py
@@ -1,0 +1,2 @@
+"""Command-line interface package for CookaReq."""
+
