+++ app/cli/__main__.py
@@ -1,0 +1,5 @@
+"""Module entry point for ``python -m app.cli``."""
+from .main import main
+
+if __name__ == "__main__":
+    main()
